Added check for pets deletion
Added check for user rights to delete own pets. If user's id does not correspond pet's owner id, you cannot delete this pet.

+++ controllers/pet/deletePet.js
@@ -6,8 +6,15 @@ const { requestError } = require("../../helpers");
 
 const deletePet = async (req, res, next) => {
   const { petId } = req.params;
+  const { id } = req.user;
 
   const pet = await UserPet.findById(petId);
+  const ownerId = pet.owner.toString();
+
+  if (ownerId !== id) {
+    throw requestError(403, "This pet is not yours");
+  }
+
   const avatarsDir = path.join(__dirname, "../", "../", "public", pet.avatarURL);
   fs.unlink(avatarsDir);
